Move converters and enums in StatsEndPoint
Move it in different folders and namespaces in StatsEndPoint
